# openapi.yaml
openapi: 3.0.3
info:
  title: School Reading & Assessment Platform API
  version: 1.0.0
  description: |
    Complete REST API for the school oral reading and AI assessment platform.

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a JWT Bearer token
    in the `Authorization: Bearer <token>` header. Access tokens expire in 15 minutes.
    Use `/auth/refresh` to obtain a new one without re-logging in.

    ## Roles
    | Role    | Access                                                        |
    |---------|---------------------------------------------------------------|
    | admin   | School structure, user management, all assignments            |
    | teacher | Their assigned sections, lesson upload, student progress view |
    | student | Their section's lessons, reading sessions, personal dashboard |

    ## Async Assessment Flow
    1. `POST /student/sessions` — creates session record
    2. Student records audio in browser via MediaRecorder API
    3. `POST /student/sessions/{id}/submit` — uploads audio (WebM/OGG)
    4. Server streams audio to Whisper STT microservice (audio never stored)
    5. Transcript sent to Ollama LLM for assessment (BullMQ job)
    6. **Socket.io** emits `assessment.complete` to student's browser
    7. `GET /student/sessions/{id}/result` — fetch the assessment result

    ## Socket.io Events
    Connect with the JWT token as a query param: `ws://server/api/v1?token=<jwt>`

...
servers:
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    SectionIdParam:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StudentIdParam:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentIdParam:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ── Reusable Schemas ──────────────────────────────────────────────────────
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-05-30T08:30:00Z"
...
    Pagination:
      type: object
      required: [total, page, limit, total_pages]
      properties:
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          nullable: true
...
    TeacherAssignment:
      type: object
      required: [id, teacher_id, teacher_name, teacher_email, section_id, section_name, class_name, is_active, assigned_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        teacher_id:
          $ref: '#/components/schemas/UUID'
        teacher_name:
          type: string
          example: Mrs. Anitha
        teacher_email:
          type: string
          example: anitha@school.edu
        section_id:
          $ref: '#/components/schemas/UUID'
        section_name:
          type: string
          example: Section A
        class_name:
          type: string
          example: Grade 5
        assigned_by_name:
          type: string
          example: Admin
        is_active:
          type: boolean
          example: true
        assigned_at:
          $ref: '#/components/schemas/Timestamp'
        unassigned_at:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Timestamp'
    CreateTeacherAssignmentRequest:
      type: object
      required: [teacher_id, section_id]
      properties:
        teacher_id:
          $ref: '#/components/schemas/UUID'
        section_id:
          $ref: '#/components/schemas/UUID'
    StudentAssignment:
      type: object
      required: [id, student_id, student_name, student_email, section_id, section_name, class_name, is_active, assigned_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        student_id:
          $ref: '#/components/schemas/UUID'
        student_name:
          type: string
          example: Arun Prakash
        student_email:
          type: string
          example: arun@school.edu
        section_id:
          $ref: '#/components/schemas/UUID'
        section_name:
          type: string
          example: Section B
        class_name:
          type: string
          example: Grade 5
        assigned_by_name:
          type: string
          example: Admin
        is_active:
          type: boolean
          example: true
        assigned_at:
          $ref: '#/components/schemas/Timestamp'
        unassigned_at:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Timestamp'
    CreateStudentAssignmentRequest:
      type: object
      required: [student_id, section_id]
      properties:
        student_id:
          $ref: '#/components/schemas/UUID'
        section_id:
          $ref: '#/components/schemas/UUID'
    MoveStudentRequest:
      type: object
      required: [new_section_id]
      properties:
        new_section_id:
          $ref: '#/components/schemas/UUID'

    # ── Lesson Schemas ────────────────────────────────────────────────────
...
    StudentProgressRow:
      type: object
      required: [student_id, student_name, lessons_completed, total_sessions, current_streak]
      properties:
        student_id:
          $ref: '#/components/schemas/UUID'
        student_name:
          type: string
          example: Arun Prakash
        lessons_completed:
          type: integer
          example: 6
        total_sessions:
          type: integer
          example: 14
        avg_overall_score:
          type: number
          format: float
          nullable: true
          example: 78.2
        avg_accuracy_score:
          type: number
          format: float
          nullable: true
          example: 84.2
        avg_fluency_score:
          type: number
          format: float
          nullable: true
          example: 71.5
        current_streak:
          type: integer
          example: 7
        last_session_date:
          type: string
          format: date
          nullable: true
          example: "2026-05-30"
        current_rank:
          type: integer
          nullable: true
          example: 3
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
    Forbidden:
      description: Authenticated but insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Resource not found
    Conflict:
      description: Resource already exists or business rule violated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: CONFLICT
            message: A section named 'Section A' already exists in Grade 5
...
paths:
...
  /admin/assignments/teachers:
    get:
      tags: [Admin — Teacher Assignments]
      summary: List teacher-section assignments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: teacher_id
          in: query
          schema:
            type: string
            format: uuid
        - name: section_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Paginated assignment list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeacherAssignment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin — Teacher Assignments]
      summary: Assign a teacher to a section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherAssignmentRequest'
            example:
              teacher_id: "550e8400-e29b-41d4-a716-446655440001"
              section_id: "550e8400-e29b-41d4-a716-446655440002"
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherAssignment'
        '400':
          description: Specified user is not a teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_ROLE
                message: The specified user does not have the teacher role
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Teacher is already actively assigned to this section
          $ref: '#/components/responses/Conflict'
  /admin/assignments/teachers/{assignmentId}:
    delete:
      tags: [Admin — Teacher Assignments]
      summary: Remove a teacher from a section
      description: Sets the assignment's is_active = false. Assignment history is preserved.
      parameters:
        - $ref: '#/components/parameters/AssignmentIdParam'
      responses:
        '204':
          description: Teacher removed from section
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — STUDENT ASSIGNMENTS
  # ═══════════════════════════════════════════════════════════════════════════
  /admin/assignments/students:
    get:
      tags: [Admin — Student Assignments]
      summary: List student-section assignments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: section_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Paginated assignment list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentAssignment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin — Student Assignments]
      summary: Assign a student to a section
      description: A student can only have one active section assignment. Returns 409 if already assigned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentAssignmentRequest'
            example:
              student_id: "550e8400-e29b-41d4-a716-446655440003"
              section_id: "550e8400-e29b-41d4-a716-446655440002"
      responses:
        '201':
          description: Student assigned to section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAssignment'
        '400':
          description: Specified user is not a student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_ROLE
                message: The specified user does not have the student role
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Student is already assigned to a section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ALREADY_ASSIGNED
                message: "Student is already assigned to Grade 5 - Section A. Use PUT /admin/assignments/students/{studentId}/move to reassign."
  /admin/assignments/students/{studentId}/move:
    put:
      tags: [Admin — Student Assignments]
      summary: Move a student to a different section
      description: |
        Deactivates the current section assignment and creates a new one.
        The old assignment row is preserved as history (is_active=false, unassigned_at set).
      parameters:
        - $ref: '#/components/parameters/StudentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveStudentRequest'
            example:
              new_section_id: "550e8400-e29b-41d4-a716-446655440005"
      responses:
        '200':
          description: Student moved to new section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAssignment'
        '400':
          description: Student is not currently assigned to any section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_ASSIGNED
                message: This student has no active section assignment
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Student is already in the target section
          $ref: '#/components/responses/Conflict'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — DASHBOARD
  # ═══════════════════════════════════════════════════════════════════════════
...
  /teacher/sections/{sectionId}/students:
    get:
      tags: [Teacher — Dashboard]
      summary: Get the progress table for all students in a section
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, avg_overall_score, lessons_completed, current_streak, last_session_date]
            default: avg_overall_score
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Partial match on student name
          schema:
            type: string
      responses:
        '200':
          description: Student progress rows
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentProgressRow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...